+++ R/x-axis.R
@@ -5,7 +5,7 @@
 ##' @param xlim xlim
 ##' @return updated tree view
 ##' @export
-##' @author guangchuang yu
+##' @author Guangchuang Yu
 xlim_tree <- function(xlim) {
     xlim_expand(xlim, panel='Tree')
 }
@@ -20,13 +20,29 @@ xlim_tree <- function(xlim) {
 ##' @return updated tree view
 ##' @importFrom ggplot2 geom_blank
 ##' @export
-##' @author guangchuang yu
+##' @author Guangchuang Yu
 xlim_expand <- function(xlim, panel) {
     structure(list(x = xlim, panel = panel), class = "facet_xlim")
 }
 
 
 
+
+##' Set y axis limits of a ggplot based on the y limits of a ggtree.
+##' This is useful for using cowplot or patchwork to properly align ggtree with other ggplot objects.
+##'
+##'
+##' @title ylim_ggtree
+##' @param ggtree ggtree object
+##' @param expand_limits amount to expand the limits
+##' @return ggplot2 object with new y limits
+##' @importFrom ggplot2 expand_scale
+##' @export
+##' @author Guangchuang Yu
+ylim_ggtree <- function(ggtree, expand_limits = expand_scale(0, 0.6)) {
+    structure(list(ggtree = ggtree, expand_limits = expand_limits),
+              class = "ylim_ggtree")
+}
 
 ##' add second x-axis for geom_range
 ##'
@@ -47,7 +63,7 @@ scale_x_range <- function() {
 ##' @param treeview treeview
 ##' @return updated treeview
 ##' @export
-##' @author guangchuang yu
+##' @author Guangchuang Yu
 revts <- function(treeview) {
     x <- treeview$data$x
     mx <- max(x)